+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## **Unreleased**
+## **Release v1.57.12 - 2026-09-17**
 **Hearthstone**:
 - Added deck highlight to Arrival of the Old Gods and Arrival of the Titans.
 - Added tracking for cost reductions from M.O.T.H.E.R.